Modularize ToGlueRow expansion logic
Move expand_to_glue_row function to a dedicated module for better organization and clarity.
Update derive macro and tests to reference the relocated expand_to_glue_row function.
Simplify a test case by reducing redundant formatting.
